chore(cleanup): cargo clippy --fix pass + delete resource_usage shim
Mechanical cleanup pass:
- cargo clippy --fix --allow-dirty applied workspace-wide. 30+
  small idiomatic fixes (deref auto, str::trim before
  split_whitespace, useless to_string(), unnecessary clones,
  redundant casts, items after test module, etc).
- Deleted app/src/resource_usage.rs compat shim from 00bd4900.
  Inlined the daemon_proto wire imports + core::process import
  at the two call sites. -1 module, -1 use re-export hop.

Also fixes a build_info.rs unnecessary-parens regression caught
by -D warnings on the cleanup pass.

cargo check --workspace --all-targets --features test-harness
clean with -D warnings; pre-existing test failures unchanged.

+++ app/src/build_info.rs
@@ -121,7 +121,7 @@ fn unix_to_ymdhm(secs: i64) -> (i32, u32, u32, u32, u32) {
     let mi = ((rem % 3600) / 60) as u32;
     let z = days + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64;
+    let doe = z - era * 146_097;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146_096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
